requirements.txt
torch==2.2.1
torchvision==0.17.1
opencv-python==4.10.0.84
pillow==10.2.0
numpy==1.24.3
scipy==1.9.3
matplotlib==3.7.2
tqdm==4.67.1
wandb==0.19.6
pyyaml==6.0.2
blobfile>=1.0.5